+++ src/lib/clearTmpDirectory.ts
@@ -5,6 +5,8 @@ export const clearTmpDirectory = async (): Promise<Option.Option<string>> => {
   try {
     await fs.emptyDir(`${__dirname}/../../tmp/working`)
     await fs.rmdir(`${__dirname}/../../tmp/working`)
+    await fs.emptyDir(`${__dirname}/../../tmp/converted`)
+    await fs.rmdir(`${__dirname}/../../tmp/converted`)
   } catch (error) {
     return Option.some(error.toString())
   }